requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
google-cloud-datastore==2.1.6
json2html==1.3.0
python-jose>=3.0.0
